# Added the ability to set custom item timings.