--- a/resources/example-configs/ktx/ktx.cfg
+++ b/resources/example-configs/ktx/ktx.cfg
@@ -121,3 +121,6 @@ set k_spec_info               1         // moreinfo (bit mask):
 set k_sayteam_to_spec         1         // send say_team to specs (0 = never, 1 = only during game, 2 = only during prewar, 3 = always)
 set allow_spec_wizard         0         // allow spectator wizards (0 = no, 1 = only when no players are on the server, 2 = only in prewar)
 set k_no_wizard_animation     0         // wizard animation (0 = on, 1 = off)
+
+// item timings
+set k_allow_custom_timings    0         // allow setting custom item timings (0 = nobody, 1 = admins only, 2 = players only, 3 = players and spectators)