There is a new command, /timing, which allows players to manage item timings
within certain limitations set by the current deathmatch mode.

Usage
-----

/timing health|ammo|armor|weapon|mega|quad|eyes|pent|all [SECONDS|default]

The workings of the command are best explained using examples.

Example 1.
Change the weapon respawn time from 30 seconds to 15 seconds (assume the
current deathmatch mode is 1):

    ] timing weapon 15

Example 2.
Change the weapon respawn time from 30 seconds to 45 seconds (assume the
current deathmatch mode is 3):

    ] timing weapon 45
    Error: item does not respawn in current deathmatch mode

Notice how the command fails in deathmatch mode 3 because in that mode
weapons don't respawn--they stay permanently on the map.

Example 3.
Fetch the current respawn time for ammo items:

    ] timing ammo
    Ammo respawn time: 15.000s

Example 4.
Set the respawn time for quad to the default value from the current
deathmatch mode:

    ] timing quad default

Example 5.
Set the respawn time for all items to the default values from the current
deathmatch mode:

    ] timing all default

Example 6.
Fetch the current respawn times for all items:

    ] timing all
    Respawn times:
     Health: 20.000s
     Ammo:   15.000s
     Armor:  20.000s
     Weapon: stay
     Mega:   20.000s
     Quad:   60.000s
     Eyes:   300.000s
     Pent:   300.000s

Example 7.
Try to set the respawn time for health to 7500 seconds:

    ] timing health 7500
    Error: duration out of range

The accepted range for SECONDS is [5,1500] (inclusive). Anything outside
of that will result in an error.

Access control
--------------

There is a new server-side cvar, allow_custom_timings, which controls who can
change item respawn times. Anybody can read the current respawn times.

    allow_custom_timings <= 0
    (default; nobody can write new timings)

    allow_custom_timings 1
    (only admins can write new timings)

    allow_custom_timings 2
    (any player, not spectator, can write new timings)

    allow_custom_timings >= 3
    (anybody can write new timings)

Miscellaneous
-------------

Whenever custom timings are in effect, the match countdown screen has a line
that reads "Custom timings" in red. If standard timings are in effect, the
line is not present.

Every time the deathmatch mode changes, custom item timings are reset to
the standard values for that mode.