@tcsabina 2021-06-16
Hi phylter.

Nice work! I've tested it, and haven't found a bug at all. Impressive ;).


Some remarks, suggestions:
1) We should list the valid range of a modified item time.
If you type `/timing health`, it should list the actual value, and the valid range. Like:
`Health respawn time: 5.000S (valid range 5-1500)`

2) Is the 5-1500 a valid range for all items? Why limit the minimum time to 5 secs?

3) Maybe consider changing the 'timing' command to 'item-timing' (or 'item_timing').

4) How 'timing all <secs>' suppose to work? I cannot find a value that is being accepted.
I see the 'timing all default' reset all timings. We need that one for sure.
But I am not sure if we need this 'timing all <secs>' possibility.
Maybe change it like this:
- remove the 'all' parameter possibility
- add 'list' as new option, so 'timing list', (or 'timing listall'), which will list the currently active times (just like 'list all' does currently)
and 'timing list' should not accept any additional parameters (so 'timing list 23' should just ignore 23)
- add 'resetall' as new parameter, (or 'timing defaultall'), which will reset all the timers (like 'timing all default' does currently)
- update the help/usage text (which is triggered when you just type 'timing', without parameter)

5) `allow_custom_timings` in the server config needs to be `k_allow_custom_timings` to be aligned with the existing naming conventions.

6) resources/example-configs/ktx/ktx.cfg needs to be updated with this new `k_allow_custom_timings` option

7) The 'Custom Timings' on the matchinfo screen is also very nice. I think we need to add there what is changed, and to what value. Something like:
'Custom Timings (h:5,w:15)'
So keep it on one line.
Or
'Custom Timings'
'h:5 w:15'
So two lines.

Whichever looks nicer on the countdown screen

8) The milliecond accuracy I think is not needed. So no need to be able to set 5.123s for an item. Integer numbers are enough.
Not sure if this simplifies anything, or makes it a bit challenging.

9) We should somehow keep these timers during map/match restarts. Currently upon the match ends, all timers are defaulted. I guess the use case would be to keep the modified timers till someone activates a usermode (1on1, 2on2, etc), or someone calls the 'timing resetall'.