@vogelca 2021-06-17
My responses to the above points:

1. We could add this. I don't feel too strongly about it either way. The permissible ranges should be large enough to cover 99% of all use cases anyway, so it shouldn't be necessary for players to keep detailed knowledge of the limits (or to deal with the additional information dump).

2. Currently the range is 5 seconds to 1500 seconds (25 minutes) for all items, but it is possible to vary the ranges per item class. It is important to understand that deathmatch modes that have weapon-stay (like DMM3) or one-time items (like DMM2) still retain those features. Those are defining characteristics of those modes and this command does not allow players to change that. If they want respawning weapons they can use DMM1 and tweak the timing if necessary. If they want weapons to stay permanently on the map, they can use DMM3. It doesn't make gameplay sense to allow the other items (like health, ammo, quad, etc.) to stay on the map permanently since players could then just sit on them and never die. I feel that 5 seconds is probably too generous already... imagine 5 second quad ;)

NOTE: Unfortunately, adding a new deathmatch mode requires MANY changes and CONSIDERABLE testing because there are literally dozens of instances throughout the code where it is assumed that there are only 5 modes, or that the modes after 3 are "special" in some way. This feature is something of a compromise that implements the requested enhancement but does not introduce such potential to break things all over the code.

3. Again, I don't feel too strongly either way. Shorter command names are generally better, provided they are not abbreviated into oblivion. If there is a real objection to the name "timing", it can be changed easily.

4. The "all" item class is there to let people fetch all the timings at once, or reset them all to defaults using one invocation of the command. For the moment the code does not permit "timing all N" (where N != "default") because some modes do not allow changing certain item times (e.g., weapon stay in DMM3). Setting all times to something other than default seemed like an unlikely operation, so I didn't spend time implementing the special case.

5. There are quite a few server-side cvars in KTX that don't start with the 'k_' prefix, but I understand the notion. It's easy enough to change.

6. Sure. No problem to add it. It does default to 0 (nobody can change item timings) if that cvar isn't defined.

7. Probably wouldn't be a bad thing, but it's not essential. It does potentially add a bunch of cryptic numbers to the countdown screen that obscure the intended message. It's not like players have much time to read it, anyway. They can always execute `/timing all` at any time if for details.

8. The millisecond accuracy (and beyond) comes for free with the use of floating-point numbers. `nextthink` is a floating-point value so any integer used must ultimately be converted. We could of course check the distance between the timing value and the nearest integer and print the integer form if that distance is less than some threshold. However, I think it's a low priority.

9. The timings are reset whenever the deathmatch mode changes. This happens when the map ends. I don't know if it's a bad thing to start out with. Custom timings should be fairly explicit.