@tcsabina 2021-06-19
My replies on your replies:
1) okay. leave it... I've checked randomly some other options/command in ezquake and I haven't found one that list the valid ranges of an option. I am sure I have seen this, but can't remember what command.
Anyhow, in the documentation we need to mention this.

2) okay.
I've played around with the 5s quad, to be honest. Loved it :).
I think this feature can make games much more fun. Fun! I don't think it would be ever used for some serious competition, so the more _crazy_ the mod is, the better. So why not a 2s health respawn time? :). Try to pop-up your enemy with a rocket, if you want to kill him ;).
But okay, leave it for now

3) okay. let's wait for some feedback on the command/option name

4) I think we discussed this briefly on Discord.
okay for now. We need to update a bit the layout of the help text, as it is confusing for the first time. It confused me, that is why I created the entry in this list at the first place.
But I accept your reasoning about it. Just make the built-in explanation/help more clear on this.

5) already done. cool!

6) no comment.

7) I still insist having this :). It will not be 'cryptic' as those who readied up have been setting this beforehand. It is just a reminder to them, the custom timing is still active.

8) How about hiding the floating point accuracy upon printing the message? I guess it is just about using %d, so it prints: weapon 15s, instead of "weapon 15.000s"

9) I think of this like a deathmatch mode. Like instagib, or lgc: when you turn them on, you want to play a couple of games with that mod on. None of these are auto disabled upon a match end/map restart. Executing explicitly a game mode, like '/1on1', shall reset these, that is okay. But a map reset should not.